backend/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==2.1.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
coverage==7.9.0
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
execnet==2.1.1
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.16.0
mypy_extensions==1.1.0
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.2.1
pytest-xdist==3.7.0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.5.0
python-slugify==8.0.4
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.0b3
starlette==0.46.2
text-unidecode==1.3
trio==0.30.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
virtualenv==20.31.2
redis>=5.0.0
